+++ src/redteam/index.ts
@@ -6,7 +6,7 @@ import * as fs from 'fs';
 import yaml from 'js-yaml';
 import logger, { getLogLevel } from '../logger';
 import type { TestCase, TestCaseWithPlugin } from '../types';
-import { checkRemoteHealth, getRemoteHealthUrl } from '../util/apiHealth';
+import { checkRemoteHealth } from '../util/apiHealth';
 import invariant from '../util/invariant';
 import { extractVariablesFromTemplates } from '../util/templates';
 import { HARM_PLUGINS, PII_PLUGINS, ALIASED_PLUGIN_MAPPINGS } from './constants';
@@ -15,7 +15,7 @@ import { extractSystemPurpose } from './extraction/purpose';
 import { Plugins } from './plugins';
 import { CustomPlugin } from './plugins/custom';
 import { redteamProviderManager } from './providers/shared';
-import { shouldGenerateRemote } from './remoteGeneration';
+import { getRemoteHealthUrl, shouldGenerateRemote } from './remoteGeneration';
 import { loadStrategy, Strategies, validateStrategies } from './strategies';
 import { DEFAULT_LANGUAGES } from './strategies/multilingual';
 import type { RedteamPluginObject, RedteamStrategyObject, SynthesizeOptions } from './types';
