+++ src/redteam/remoteGeneration.ts
@@ -21,6 +21,28 @@ export function neverGenerateRemote(): boolean {
   return getEnvBool('PROMPTFOO_DISABLE_REDTEAM_REMOTE_GENERATION');
 }
 
+/**
+ * Gets the URL for checking remote API health based on configuration.
+ * @returns The health check URL, or null if remote generation is disabled.
+ */
+export function getRemoteHealthUrl(): string | null {
+  if (neverGenerateRemote()) {
+    return null;
+  }
+
+  const envUrl = getEnvString('PROMPTFOO_REMOTE_GENERATION_URL');
+  if (envUrl) {
+    return envUrl.replace(/\/task$/, '/health');
+  }
+
+  const cloudConfig = new CloudConfig();
+  if (cloudConfig.isEnabled()) {
+    return `${cloudConfig.getApiHost()}/health`;
+  }
+
+  return 'https://api.promptfoo.app/health';
+}
+
 export function shouldGenerateRemote(): boolean {
   // Generate remotely when the user has not disabled it and does not have an OpenAI key.
   return (!neverGenerateRemote() && !getEnvString('OPENAI_API_KEY')) || (cliState.remote ?? false);
