+++ test/util/apiHealth.test.ts
@@ -1,4 +1,5 @@
-import { checkRemoteHealth, getRemoteHealthUrl } from '../../src/util/apiHealth';
+import { getRemoteHealthUrl } from '../../src/redteam/remoteGeneration';
+import { checkRemoteHealth } from '../../src/util/apiHealth';
 
 const mockedFetch = jest.mocked(jest.fn());
 global.fetch = mockedFetch;
